requirements.txt
aws-cdk.aws-apigateway==1.115.0
aws-cdk.aws-codebuild==1.115.0
aws-cdk.aws-codecommit==1.115.0
aws-cdk.aws-codedeploy==1.115.0
aws-cdk.aws-codepipeline==1.115.0
aws-cdk.aws-codepipeline-actions==1.115.0
aws-cdk.aws-ecr==1.115.0
aws-cdk.core==1.115.0
aws-cdk.pipelines==1.115.0
botocore==1.20.77
aws-cdk.aws-ecs-patterns==1.115.0
aws-cdk.aws-rds==1.115.0
aws-cdk.aws-lambda-python==1.115.0
aws-cdk.cloudformation-include==1.115.0
